requirements.txt
Django==4.2.8
djangorestframework==3.14.0
Pillow==10.1.0
numpy==1.21.2
google-cloud-vision==3.5.0